Refactored XML utils
Refactored XML utils
