+++ src/main/java/eu/ortlepp/blogbuilder/action/Build.java
@@ -2,13 +2,13 @@ package eu.ortlepp.blogbuilder.action;
 
 import eu.ortlepp.blogbuilder.model.Document;
 import eu.ortlepp.blogbuilder.util.Cleaner;
-import eu.ortlepp.blogbuilder.util.FeedCreator;
 import eu.ortlepp.blogbuilder.util.ResourceCopy;
 import eu.ortlepp.blogbuilder.util.Scanner;
-import eu.ortlepp.blogbuilder.util.SitemapCreator;
 import eu.ortlepp.blogbuilder.util.Writer;
 import eu.ortlepp.blogbuilder.util.config.Config;
 import eu.ortlepp.blogbuilder.util.config.Directories;
+import eu.ortlepp.blogbuilder.util.xml.FeedCreator;
+import eu.ortlepp.blogbuilder.util.xml.SitemapCreator;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
